+++ packages/ui/src/components/ErrorPresenter.svelte
@@ -14,12 +14,12 @@
 // limitations under the License.
 -->
 <script lang="ts">
-  import Tooltip from './Tooltip.svelte'
+  import { tooltip } from '../tooltips'
   import ErrorPopup from './ErrorPopup.svelte'
   export let error: any
 </script>
 
-<Tooltip component={ErrorPopup} props={{ error: error }}>
+<div use:tooltip={{ component: ErrorPopup, props: { error: error } }}>
   <svg width="20" height="20" viewBox="0 0 20 20" fill="none" xmlns="http://www.w3.org/2000/svg">
     <circle cx="10" cy="10" r="7.275" stroke="#EE7A7A" stroke-width="1.2" />
     <path
@@ -29,4 +29,4 @@
       stroke-width="0.2"
     />
   </svg>
-</Tooltip>
+</div>